adds a retriable bootstrap prebuilt rules with a real package
