+++ x-pack/test/security_solution_api_integration/test_suites/detections_response/rules_management/prebuilt_rules/common/configs/edge_cases/ess_trial_license.config.ts
@@ -19,6 +19,10 @@ export default async function ({ readConfigFile }: FtrConfigProviderContext) {
       reportName:
         'Rules Management - Prebuilt Rules (Common) Integration Tests - ESS Trial License',
     },
+    mochaOpts: {
+      ...functionalConfig.get('mochaOpts'),
+      timeout: 60000 * 10, // 10 minutes
+    },
   };
 
   return testConfig;
