+++ x-pack/test/security_solution_api_integration/test_suites/detections_response/rules_management/prebuilt_rules/common/prebuilt_rules_package/air_gapped/index.ts
@@ -10,6 +10,7 @@ import { FtrProviderContext } from '../../../../../../../ftr_provider_context';
 export default ({ loadTestFile }: FtrProviderContext): void => {
   describe('Air-gapped environment with pre-bundled packages', function () {
     this.tags('skipFIPS');
+    loadTestFile(require.resolve('./bootstrap_prebuilt_rules'));
     loadTestFile(require.resolve('./install_bundled_package'));
     loadTestFile(require.resolve('./prerelease_packages'));
   });
