+++ x-pack/test/security_solution_api_integration/test_suites/detections_response/rules_management/prebuilt_rules/common/prebuilt_rules_package/install_package_from_epr.ts
@@ -5,6 +5,10 @@
  * 2.0.
  */
 import expect from 'expect';
+import {
+  ENDPOINT_PACKAGE_NAME,
+  PREBUILT_RULES_PACKAGE_NAME,
+} from '@kbn/security-solution-plugin/common/detection_engine/constants';
 import { FtrProviderContext } from '../../../../../../ftr_provider_context';
 import {
   getPrebuiltRulesAndTimelinesStatus,
@@ -13,7 +17,10 @@ import {
 import { deleteAllRules } from '../../../../../../../common/utils/security_solution';
 import { deleteAllPrebuiltRuleAssets } from '../../../../utils/rules/prebuilt_rules/delete_all_prebuilt_rule_assets';
 import { deleteAllTimelines } from '../../../../utils/rules/prebuilt_rules/delete_all_timelines';
-import { deletePrebuiltRulesFleetPackage } from '../../../../utils/rules/prebuilt_rules/delete_fleet_packages';
+import {
+  deleteEndpointFleetPackage,
+  deletePrebuiltRulesFleetPackage,
+} from '../../../../utils/rules/prebuilt_rules/delete_fleet_packages';
 import { installPrebuiltRulesFleetPackage } from '../../../../utils/rules/prebuilt_rules/install_prebuilt_rules_fleet_package';
 
 export default ({ getService }: FtrProviderContext): void => {
@@ -21,13 +28,43 @@ export default ({ getService }: FtrProviderContext): void => {
   const supertest = getService('supertest');
   const log = getService('log');
   const retryService = getService('retry');
+  const securitySolutionApi = getService('securitySolutionApi');
 
   describe('@ess @serverless @skipInServerlessMKI Install prebuilt rules from EPR', () => {
     beforeEach(async () => {
-      await deletePrebuiltRulesFleetPackage({ supertest, es, log, retryService });
       await deleteAllRules(supertest, log);
       await deleteAllTimelines(es, log);
       await deleteAllPrebuiltRuleAssets(es, log);
+    });
+
+    it('bootstraps prebuilt rules by installing required packages from EPR', async () => {
+      await retryService.tryWithRetries(
+        'bootstrapPrebuiltRules',
+        async () => {
+          await deletePrebuiltRulesFleetPackage({ supertest, es, log, retryService });
+          await deleteEndpointFleetPackage({ supertest, es, log, retryService });
+
+          const { body } = await securitySolutionApi.bootstrapPrebuiltRules().expect(200);
+
+          expect(body).toMatchObject({
+            packages: expect.arrayContaining([
+              expect.objectContaining({
+                name: PREBUILT_RULES_PACKAGE_NAME,
+                status: 'installed',
+              }),
+              expect.objectContaining({
+                name: ENDPOINT_PACKAGE_NAME,
+                status: 'installed',
+              }),
+            ]),
+          });
+        },
+        {
+          retryCount: 10, // 10s delay * 120 reties = 20 mins
+          retryDelay: 5000,
+          timeout: 60000 * 10, // total timeout applied to all attempts altogether, 10 mins
+        }
+      );
     });
 
     /**
@@ -35,6 +72,9 @@ export default ({ getService }: FtrProviderContext): void => {
      * package storage and checks that they are installed.
      */
     it('should install prebuilt rules from the package storage', async () => {
+      await deletePrebuiltRulesFleetPackage({ supertest, es, log, retryService });
+      await deleteEndpointFleetPackage({ supertest, es, log, retryService });
+
       // Verify that status is empty before package installation
       const statusBeforePackageInstallation = await getPrebuiltRulesAndTimelinesStatus(
         es,
